fix: show version information for plugin and loader
