+++ packages/webpack-cli/lib/groups/HelpGroup.js
@@ -1,5 +1,6 @@
 const chalk = require('chalk');
 const { core, commands } = require('../utils/cli-flags');
+const { defaultCommands } = require('../utils/commands');
 const commandLineUsage = require('command-line-usage');
 
 class HelpGroup {
@@ -49,7 +50,7 @@ class HelpGroup {
         if (externalPkg && commandsUsed.length === 1 && invalidArgs.length === 0) {
             try {
                 if (commandsUsed.includes(externalPkg.name)) {
-                    const { name, version } = require(`@webpack-cli/${externalPkg.name}/package.json`);
+                    const { name, version } = require(`@webpack-cli/${defaultCommands[externalPkg.name]}/package.json`);
                     process.stdout.write(`\n${name} ${version}`);
                 } else {
                     const { name, version } = require(`${externalPkg.name}/package.json`);
