+++ packages/webpack-cli/lib/utils/commands.js
@@ -1,5 +1,14 @@
 const { commands } = require('./cli-flags');
 
+const defaultCommands = {
+    init: 'init',
+    loader: 'generate-loader',
+    plugin: 'generate-plugin',
+    info: 'info',
+    migrate: 'migrate',
+    serve: 'serve',
+};
+
 // Contains an array of strings with commands and their aliases that the cli supports
 const names = commands
     .map(({ alias, name }) => {
@@ -10,4 +19,7 @@ const names = commands
     })
     .reduce((arr, val) => arr.concat(val), []);
 
-module.exports = { names };
+module.exports = {
+    defaultCommands,
+    names,
+};
